Remove specification concept; replace with generic Attribute<E>
After some analysis, it seemed apparent that the AttributeSpecification, DataSetSpecification, etc., classes were a complication of the attribute concept. Attributes, basically a key-value pair, can be sufficiently implemented using a Java class for the value type and an overriden isValid() method for validation. The previous strategy was to use a series of ...Specification classes, one for every type of Attribute and DataSet. The new strategy uses generics (Attribute<E>), which is more simple and reduces class clutter.

This commit fixes most usage in the framework package. Further work will have to clear out usage from the inductive, knowledge, and learners packages.